kommentare und details

+++ src/BMIRechner.java
@@ -1,19 +1,25 @@
 public class BMIRechner {
 
-    private float mGewicht= 65.1f;
-    private float mGroesse=1.6f;
+    private double mGewicht = 65.1;
+    private double mGroesse = 1.6;
 
-    public double bmiBerechnen(){
+    /**
+     * BMI wird berechnet mit hard gecodeten Werten, Ergebnis wird returned
+     */
+    public double bmiBerechnen() {
         double bmi = (this.mGewicht/ Math.pow(this.mGroesse, 2));
         return bmi;
     }
 
-    public String diagnose(Double bmi){
-        if (bmi<=18.5){
+    /**
+     * anhand des uebergebenen BMI-Wertes wird die Diagnose erstellt
+     */
+    public String diagnose(Double bmi) {
+        if (bmi <= 18.5) {
             return "Untergewicht";
-        } else if (bmi>=25){
+        } else if (bmi >= 25) {
             return "Übergewicht";
-        } else{
+        } else {
             return "Normalgewicht";
         }    
     }
